refactor(web): Split out Factory Imports section into it's own area
