+++ web/src/components.d.ts
@@ -12,9 +12,11 @@ declare module 'vue' {
     DroneCalculator: typeof import('./components/planner/satisfaction/calculator/DroneCalculator.vue')['default']
     ExportCalculator: typeof import('./components/planner/satisfaction/calculator/ExportCalculator.vue')['default']
     FactoryDebug: typeof import('./components/FactoryDebug.vue')['default']
+    FactoryImports: typeof import('./components/planner/imports/FactoryImports.vue')['default']
     FactoryNode: typeof import('./components/graph/FactoryNode.vue')['default']
     GameAsset: typeof import('./components/GameAsset.vue')['default']
     Graph: typeof import('./components/graph/Graph.vue')['default']
+    Imports: typeof import('./components/planner/imports/Imports.vue')['default']
     Introduction: typeof import('./components/planner/Introduction.vue')['default']
     JoinDiscord: typeof import('./components/JoinDiscord.vue')['default']
     KoFi: typeof import('./components/ko-fi.vue')['default']
@@ -24,7 +26,7 @@ declare module 'vue' {
     Placeholder: typeof import('./components/planner/Placeholders/Placeholder.vue')['default']
     Planner: typeof import('./components/planner/Planner.vue')['default']
     PlannerFactory: typeof import('./components/planner/PlannerFactory.vue')['default']
-    PlannerFactoryImports: typeof import('./components/planner/PlannerFactoryImports.vue')['default']
+    PlannerFactoryImports: typeof import('./components/planner/imports/PlannerFactoryImports.vue')['default']
     PlannerFactoryList: typeof import('./components/planner/PlannerFactoryList.vue')['default']
     PlannerFactoryNotes: typeof import('./components/planner/PlannerFactoryNotes.vue')['default']
     PlannerFactoryPlaceholderList: typeof import('./components/planner/Placeholders/PlannerFactoryPlaceholderList.vue')['default']
@@ -37,6 +39,7 @@ declare module 'vue' {
     PowerProducer: typeof import('./components/planner/products/PowerProducer.vue')['default']
     Product: typeof import('./components/planner/products/Product.vue')['default']
     ProductsAndPower: typeof import('./components/planner/products/ProductsAndPower.vue')['default']
+    RawResources: typeof import('./components/planner/imports/RawResources.vue')['default']
     RecipeItem: typeof import('./components/recipes/RecipeItem.vue')['default']
     Recipes: typeof import('./components/recipes/Recipes.vue')['default']
     RecipeSearchItem: typeof import('./components/recipes/RecipeSearchItem.vue')['default']
