+++ web/src/components/planner/imports/FactoryImports.vue
@@ -1,0 +1,29 @@
+<template>
+  <div>
+    <h1 class="text-h5 mb-4">
+      <i class="fas fa-arrow-to-right" />
+      <span class="ml-3">Imports</span>
+    </h1>
+    <p v-show="helpText" class="text-body-2 mb-4">
+      <i class="fas fa-info-circle" /> Imports are the resources needed to produce the factory's products and ensure its satisfaction. To set up imports, you select another factory and choose one of its outputs. This creates a "request" for that output. The selected factory must fulfill this request, and you'll be notified if it cannot meet the demand. All available outputs are listed in the Outputs section of the factory you choose.
+    </p>
+    <div v-if="Object.keys(factory.rawResources).length > 0 || Object.keys(factory.parts).length > 0">
+      <raw-resources :factory="factory" />
+      <imports :factory="factory" :help-text="helpText" />
+    </div>
+    <p v-else class="text-body-1">Awaiting product selection.</p>
+  </div>
+</template>
+
+<script setup lang="ts">
+  import { defineProps } from 'vue'
+  import { Factory } from '@/interfaces/planner/FactoryInterface'
+  import RawResources from '@/components/planner/imports/RawResources.vue'
+  import Imports from '@/components/planner/imports/Imports.vue'
+
+  defineProps<{
+    factory: Factory;
+    helpText: boolean;
+  }>()
+
+</script>
